pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "hephaestus"
version = "0.1.0"
description = "Standalone developer toolkit for orchestrating quality and refactoring workflows."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Hephaestus Team", email = "opensource@hephaestus.dev" }
]
license = { file = "LICENSE" }
keywords = ["refactoring", "quality", "tooling", "cli"]
dependencies = [
  "typer[all]>=0.12.3",
  "rich>=14.1.0",
  "libcst>=1.2",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "tomli; python_version < '3.11'"
]

[project.optional-dependencies]
qa = [
  "coverage[toml]>=7.4",
  "pytest>=8.2"
]

[project.scripts]
hephaestus = "hephaestus.cli:app"

[tool.hephaestus.toolkit]
default_config = "hephaestus-toolkit/refactoring/config/refactor.config.yaml"
workspace_root = "."

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["src"]
filterwarnings = [
  "error"
]
